requirements.txt
django-bootstrap-v5
django-bootstrap-icons